Remove global catalog
