+++ internal/catalog/builtin/lighthouse.star
