+++ internal/catalog/builtin/teku.star
